*Edited Maze.java file to add new feature
Removed isWalkable as it conflicted with the edited

Added and statements to each direction in takeStep so that if the
"character" walks into a wall and has to turn around they can walk on 1s
as well as 2s

Edited moveNorth, moveEast, moveWest, and moveSouth so that when the
"character" walks on a 2 tile it will turn the previous tile into a 3
which will be displayed as a space. This will give the appearance that
the wrong paths are being deleted and since they are filled with 3s they
can no longer be walked on blocking off that path.

As a result the maze path now looks cleaner and shows the path to the
exit without showing any of the "bad paths"